+++ app/features/quiz/types/index.ts
@@ -1,0 +1,5 @@
+export interface QuizData {
+	word: string;
+	translation: string;
+	frequency: number;
+}
